+++ AddOns/Xfa/Source/wali/util/base64.hpp
@@ -5,7 +5,7 @@
 
 namespace wali {
     namespace util {
-        std::string base64_encode(unsigned char const* , unsigned int len);
+        std::string base64_encode(unsigned char const* , size_t len);
         std::string base64_decode(std::string const& s);
     }
 }
